Clean up and add suggested demo script (#36)
* Factor out IngredientsListEditor

* Don't use nested form as it breaks prerendering

* Simplify

* Demo script
